requirements.txt
lxml==3.6.4
Pillow==3.4.2
python-docx==0.8.6
